+++ app/src/main/java/gortea/jgmax/animalstorage/ui/viewmodel/interfaces/AnimalDataToItemConverter.kt
@@ -3,5 +3,5 @@ package gortea.jgmax.animalstorage.ui.viewmodel.interfaces
 import gortea.jgmax.animalstorage.data.model.AnimalItem
 
 interface AnimalDataToItemConverter {
-    fun dataToItem(name: String, age: Int, breed: String, id: Long = 0): AnimalItem
+    fun dataToItem(name: String, age: Int, breed: String, id: Int = 0): AnimalItem
 }
